server: Move platform dependent code to separate module.
This patch moves ServiceDeps and ContextDeps along side the whole crypto
module and most the buffer module into a new module called "platform"
That module contains everything server user must implement for
themselves.

The remaining part of buffer module was for handling requests/responses
so it was renamed to req_resp.

This patch does not modify any business logic in anyway.
