+++ server/src/lib.rs
@@ -1,9 +1,9 @@
 #![cfg_attr(not(test), no_std)]
 #![forbid(unsafe_code)]
 
-pub mod buffer;
-pub mod crypto;
 mod handler;
+pub mod platform;
+pub mod req_resp;
 pub mod service;
 #[cfg(test)]
 mod tests;
