+++ server/src/platform/crypto.rs
